clean up network type file and enforce NodeType
