+++ common/v2/types/network.ts
@@ -1,48 +1,10 @@
 import { Subtract } from 'utility-types';
-import { BlockExplorer } from './blockExplorer';
-import { SecureWalletName, InsecureWalletName } from './wallet';
-import { GasPrice } from './gas';
 
-export type NetworkId =
-  | 'ETH'
-  | 'Ropsten'
-  | 'Kovan'
-  | 'Rinkeby'
-  | 'Goerli'
-  | 'ETC'
-  | 'RSK'
-  | 'AKA'
-  | 'AQUA'
-  | 'ARTIS_SIGMA1'
-  | 'ARTIS_TAU1'
-  | 'ATH'
-  | 'CLO'
-  | 'DEXON'
-  | 'EGEM'
-  | 'ELLA'
-  | 'EOSC'
-  | 'ESN'
-  | 'ETI'
-  | 'ETHO'
-  | 'ETSC'
-  | 'EXP'
-  | 'Gangnam'
-  | 'GO'
-  | 'GO_TESTNET'
-  | 'METADIUM'
-  | 'MIX'
-  | 'MUSIC'
-  | 'PIRL'
-  | 'POA'
-  | 'REOSC'
-  | 'RSK_TESTNET'
-  | 'SOLIDUM'
-  | 'THUNDERCORE'
-  | 'TOMO'
-  | 'UBQ'
-  | 'WEB'
-  | 'AUX'
-  | 'ASK';
+import { NodeOptions } from './node';
+import { BlockExplorer } from './blockExplorer';
+import { GasPrice } from './gas';
+import { DPathFormats } from './dPath';
+import { NetworkId } from './networkId';
 
 export interface Network {
   id: NetworkId;
@@ -85,70 +47,3 @@ interface NetworkUnusedLegacyProps {
 
 export type NetworkLegacy = Subtract<ExtendedNetwork, NetworkPropsMissingInLegacy> &
   NetworkUnusedLegacyProps;
-
-export interface NodeOptions {
-  name: string;
-  type?: 'rpc' | 'etherscan' | 'infura' | 'web3' | 'myccustom';
-  service: string;
-  url: string;
-  isCustom?: boolean;
-  isAuto?: boolean;
-  network?: string;
-  hidden?: boolean;
-}
-
-export interface ExtendedNodeOptions extends NodeOptions {
-  uuid: string;
-}
-
-export interface CustomNodeConfig {
-  id: string;
-  isCustom: true;
-  isAuto?: undefined;
-  name: string;
-  service: 'your custom node';
-  url: string;
-  network: string;
-  auth?: {
-    username: string;
-    password: string;
-  };
-}
-
-export interface StaticNodeConfig {
-  id: string;
-  isCustom: false;
-  isAuto?: boolean;
-  network: NetworkId;
-  service: string;
-  hidden?: boolean;
-}
-
-export interface DerivationPathOptions {
-  name: string;
-  derivationPath: string;
-  active: boolean;
-}
-
-export interface ExtendedDerivationPathOptions extends DerivationPathOptions {
-  uuid: string;
-}
-
-export interface DPathFormats {
-  default?: DPath;
-  trezor?: DPath;
-  safeTmini?: DPath;
-  ledgerNanoS?: DPath;
-  mnemonicPhrase: DPath;
-}
-
-export interface DPath {
-  label: string;
-  value: string; // TODO determine method for more precise typing for path
-}
-
-export type DPathFormat =
-  | SecureWalletName.TREZOR
-  | SecureWalletName.SAFE_T
-  | SecureWalletName.LEDGER_NANO_S
-  | InsecureWalletName.MNEMONIC_PHRASE;
