+++ common/v2/types/networkId.ts
@@ -1,0 +1,40 @@
+export type NetworkId =
+  | 'ETH'
+  | 'Ropsten'
+  | 'Kovan'
+  | 'Rinkeby'
+  | 'Goerli'
+  | 'ETC'
+  | 'RSK'
+  | 'AKA'
+  | 'AQUA'
+  | 'ARTIS_SIGMA1'
+  | 'ARTIS_TAU1'
+  | 'ATH'
+  | 'CLO'
+  | 'DEXON'
+  | 'EGEM'
+  | 'ELLA'
+  | 'EOSC'
+  | 'ESN'
+  | 'ETI'
+  | 'ETHO'
+  | 'ETSC'
+  | 'EXP'
+  | 'Gangnam'
+  | 'GO'
+  | 'GO_TESTNET'
+  | 'METADIUM'
+  | 'MIX'
+  | 'MUSIC'
+  | 'PIRL'
+  | 'POA'
+  | 'REOSC'
+  | 'RSK_TESTNET'
+  | 'SOLIDUM'
+  | 'THUNDERCORE'
+  | 'TOMO'
+  | 'UBQ'
+  | 'WEB'
+  | 'AUX'
+  | 'ASK';
